AckPacket class
