+++ src/AcknowledgePacket.cpp
@@ -1,7 +1,5 @@
-#include "DataPacket.hpp"
-#include "HardwareSerial.h"
-#include "NodeRegistry.hpp"
 #include "AcknowledgePacket.hpp"
+#include "NodeRegistry.hpp"
 #include "Protocol.hpp"
 #include "config.hpp"
 #include "utils.hpp"
@@ -12,38 +10,14 @@
 #include <cstdlib>
 #include <WiFi.h>
 
-DataPacket::DataPacket(){
-}
-
-
-DataPacket::DataPacket(const std::string& message, const uint8_t dest[6]){
-    this->type = PacketType::DAT;
-
-    uint8_t mac_bytes[6];
-    macStringToBytes(WiFi.macAddress(), mac_bytes);
-    memcpy(this->src, mac_bytes, 6);
-    memcpy(this->dest, dest, 6);
-
-    size_t len = std::min(message.size(), (size_t)DATA_MESSAGE_SIZE);
-    memcpy(this->msg, message.data(), len);
-    if (len < DATA_MESSAGE_SIZE) {
-        memset(this->msg + len, 0, DATA_MESSAGE_SIZE - len);
-    }
-
-    this->ttl = DATA_TTL;
-    this->pkt_id = generatePktId(mac_bytes);
-    this->chs = this->checksum();
-}
-
-
 // Little-Endian
-size_t DataPacket::serialize(uint8_t* buffer, size_t buffer_size) const {
+size_t AcknowledgePacket::serialize(uint8_t* buffer, size_t buffer_size) const {
   // PacketType type;
   // uint8_t src[6];
   // uint16_t pkt_id;
+  // uint16_t ack_pkt_id;
   // uint8_t dest[6];
   // uint8_t ttl;
-  // uint8_t msg[20];
   // uint8_t chs;
   
   size_t offset = 0;
@@ -58,6 +32,9 @@ size_t DataPacket::serialize(uint8_t* buffer, size_t buffer_size) const {
   if(offset + 2 > buffer_size) return offset;
   writeLE16(buffer, offset, this->pkt_id);
 
+  if(offset + 2 > buffer_size) return offset;
+  writeLE16(buffer, offset, this->ack_pkt_id);
+
   if(offset + 6 > buffer_size) return offset;
   memcpy(buffer + offset, this->dest, 6);
   offset += 6;
@@ -65,21 +42,17 @@ size_t DataPacket::serialize(uint8_t* buffer, size_t buffer_size) const {
   if(offset + 1 > buffer_size) return offset;
   buffer[offset++] = this->ttl;
 
-  if(offset + DATA_MESSAGE_SIZE > buffer_size) return offset;
-  memcpy(buffer + offset, this->msg, DATA_MESSAGE_SIZE);
-  offset += DATA_MESSAGE_SIZE;
-
   if(offset + 1 > buffer_size) return offset;
   buffer[offset++] = this->chs;
 
   return offset;
 }
 
-uint8_t DataPacket::checksum() const{
+uint8_t AcknowledgePacket::checksum() const{
   // Could be done diretly on field to save memory
-  // Size of the packet is 37
+  // Size of the packet is 19
   uint8_t cs = 0;
-  uint8_t bytes[DATA_MESSAGE_SIZE + 17];
+  uint8_t bytes[19];
   size_t len = this->serialize(bytes, sizeof(bytes));
   for(size_t i = 0; i < len - 1; i++){
     cs ^= bytes[i];
@@ -87,8 +60,8 @@ uint8_t DataPacket::checksum() const{
   return cs;
 }
 
-bool DataPacket::deserializeFields(const uint8_t* buffer, size_t len){
-  if(len < DATA_MESSAGE_SIZE + 17) return false;
+bool AcknowledgePacket::deserializeFields(const uint8_t* buffer, size_t len){
+  if(len < 19) return false;
 
   size_t offset = 0;
 
@@ -104,6 +77,10 @@ bool DataPacket::deserializeFields(const uint8_t* buffer, size_t len){
   this->pkt_id = readLE16(buffer, offset);
   offset += 2;
 
+  if(offset + 2 > len) return false;
+  this->ack_pkt_id = readLE16(buffer, offset);
+  offset += 2;
+
   if(offset + 6 > len) return false;
   memcpy(this->dest, buffer + offset, 6);
   offset += 6;
@@ -111,17 +88,32 @@ bool DataPacket::deserializeFields(const uint8_t* buffer, size_t len){
   if(offset + 1 > len) return false;
   this->ttl = buffer[offset++];
 
-  if(offset + DATA_MESSAGE_SIZE > len) return false;
-  memcpy(this->msg, buffer + offset, DATA_MESSAGE_SIZE);
-  offset += DATA_MESSAGE_SIZE;
-
   if(offset + 1 > len) return false;
   this->chs = buffer[offset++];
 
   return true;
 }
 
-void DataPacket::handle() {
+AcknowledgePacket::AcknowledgePacket(const uint8_t dest[6], const uint16_t packet_id){
+    this->type = PacketType::ACK;
+
+    uint8_t mac_bytes[6];
+    macStringToBytes(WiFi.macAddress(), mac_bytes);
+    memcpy(this->src, mac_bytes, 6);
+    memcpy(this->dest, dest, 6);
+
+    this->ttl = ACK_TTL;
+    this->pkt_id = generatePktId(this->src);
+    this->ack_pkt_id = packet_id;
+    this->chs = this->checksum();
+}
+
+AcknowledgePacket::AcknowledgePacket(){}
+
+void AcknowledgePacket::handle() {
+  //TODO: hadling logic here
+  // Serial.printf("AcknowledgePacket handling here\n");
+  // Serial.println(macBytesToString(this->src));
   this->ttl = this->ttl - 1;
   this->chs = this->checksum();
   if(this->ttl > 0){
@@ -135,14 +127,10 @@ void DataPacket::handle() {
         }
     }
     if(equal){
-      Serial.print("Message: ");
-      Serial.println((char*)this->msg);
-      Serial.println("Sending ACK");
-      AcknowledgePacket ack_pkt(this->src, this->pkt_id);
-      sendPacket(this->src, ack_pkt);
+      Serial.print("ACK FROM: ");
+      Serial.println(macBytesToString(this->src));
     } else {
       if(NodeRegistry::instance().peerExists(this->dest)){
-        Serial.println("Sending to dest!");
         sendPacket(this->dest, *this);
       }else{
         sendPacket(NodeRegistry::instance().getMostRecentNode().data(), *this);
